requirements.txt
scikit-learn
numpy
matplotlib
joblib
Pillow